Created basic MVC implementation
- Added front controller
- Added Router and Regex Route implementations
- Added abstract RESTful controller implementation

+++ library/mwop/Controller/Restful.php
@@ -2,31 +2,77 @@
 namespace mwop\Controller;
 
 use mwop\Stdlib\Dispatchable,
+    mwop\Stdlib\Entity,
+    mwop\Stdlib\Resource,
     Fig\Request,
     Fig\Response,
     Zend\Http\Request as HttpRequest,
     Zend\Http\Response as HttpResponse,
     Zend\EventManager\EventCollection,
-    Zend\EventManager\EventManager;
+    Zend\EventManager\EventManager,
+    Zend\Filter\InputFilter;
 
 abstract class Restful implements Dispatchable
 {
     protected $request;
     protected $response;
     protected $events;
+    protected $resource;
 
-    abstract public function getList();
-    abstract public function get($id);
-    abstract public function create($data);
-    abstract public function update($id, $data);
-    abstract public function delete($id);
+    abstract public function resource(Resource $resource = null);
+
+    public function getList()
+    {
+        $entities = $this->resource()->getAll()->toArray();
+        return array( 
+            'entities' => array_values($entities),
+        );
+    }
+
+    public function get($id)
+    {
+        return array(
+            'entity' => $this->resource()->get($id)->toArray(),
+        );
+    }
+
+    public function create($data)
+    {
+        $entity = $this->resource()->create($data);
+        if ($entity instanceof Entity) {
+            return array(
+                'entity'  => $entity->toArray(),
+                'success' => true,
+            );
+        }
+        return array(
+            'success' => false,
+            'errors'  => $entity->getMessages(),
+        );
+    }
+
+    public function update($id, $data)
+    {
+        return array(
+            'entity' => $this->resource()->update($id, $data)->toArray(),
+        );
+    }
+
+    public function delete($id)
+    {
+        return array(
+            'success' => $this->resource()->delete($id),
+        );
+    }
 
     public function events(EventCollection $events = null)
     {
         if (null !== $events) {
             $this->events = $events;
         } elseif (null === $this->events) {
-            $this->events = new EventManager(array(__CLASS__, get_called_class()));
+            $this->events = new EventManager(array(
+                'mwop\Stdlib\Dispatchable', __CLASS__, get_called_class()
+            ));
         }
         return $this->events;
     }
@@ -51,32 +97,44 @@ abstract class Restful implements Dispatchable
         $this->setRequest($request)
              ->setResponse($response);
 
-        switch (strtolower($request->getMethod())) {
-            case 'get':
-                if (null !== $id = $request->getMetadata('id')) {
-                    $return = $this->get($id);
+        if ($action = $request->getMetadata('action', false)) {
+            // Handle arbitrary methods, ending in Action
+            $method = $this->getMethodFromAction($action);
+            if (!method_exists($this, $method)) {
+                $this->prepare404();
+                return $response;
+            }
+            $return = $this->$method();
+        } else {
+            // RESTful methods
+            switch (strtolower($request->getMethod())) {
+                case 'get':
+                    if (null !== $id = $request->getMetadata('id')) {
+                        $return = $this->get($id);
+                        break;
+                    }
+                    $return = $this->getList();
                     break;
-                }
-                $return = $this->getAll();
-                break;
-            case 'post':
-                $return = $this->post($request->post());
-                break;
-            case 'put':
-                if (null === $id = $request->getMetadata('id')) {
-                    throw new \DomainException('Missing identifier');
-                }
-                $params = $request->getContent();
-                $return = $this->put($id, $params);
-                break;
-            case 'delete':
-                if (null === $id = $request->getMetadata('id')) {
-                    throw new \DomainException('Missing identifier');
-                }
-                $return = $this->delete($id);
-                break;
-            default:
-                throw new \DomainException('Invalid HTTP method!');
+                case 'post':
+                    $return = $this->create($request->post()->toArray());
+                    break;
+                case 'put':
+                    if (null === $id = $request->getMetadata('id')) {
+                        throw new \DomainException('Missing identifier');
+                    }
+                    $params = $request->getContent();
+                    $params = parse_str($params);
+                    $return = $this->update($id, $params);
+                    break;
+                case 'delete':
+                    if (null === $id = $request->getMetadata('id')) {
+                        throw new \DomainException('Missing identifier');
+                    }
+                    $return = $this->delete($id);
+                    break;
+                default:
+                    throw new \DomainException('Invalid HTTP method!');
+            }
         }
 
         // Emit post-dispatch signal, passing:
@@ -144,5 +202,12 @@ abstract class Restful implements Dispatchable
             $this->setResponse(new HttpResponse());
         }
         return $this->response;
+    }
+
+    protected function getMethodFromAction($action)
+    {
+        $method = str_replace(array('-', '.', '_'), ' ', $action);
+        $method = ucwords($method);
+        return str_replace(' ', '', $method) . 'Action';
     }
 }
